requirements.txt
streamlit
pandas
plotly
sqlalchemy
psycopg2-binary
python-dotenv
pyarrow
pyspark